feature[portfolio-v2] : getColor bubble utils
